Engine: PoolMgr & SceneMgr
This commit adds more features to the Bird Soul Engine.

+PoolMgr: Used as GameObjects cache, allowing the preload of multiple
instances of a prefab  via BSEngineLoader script

+SceneMgr, Scenes and Subscenes: Now App states loads an initial scene.
With SceneMgr youa load more scenes as "Subscenes" to a state's intial
scene. This manager will support async operations, eventually...

All the structure needed to begin to create scenes with the engine is
ready. The current state of the repo is a test benchmark with dummy
scenes.